AIAutomaticDrawing/requirements.txt

PyQt5
pandas
matplotlib
pandasai
pandasai-litellm
Pillow